Module aws_sdk_chimesdkidentity::error

Modules

bad_request_exception

See BadRequestException

conflict_exception

See ConflictException

forbidden_exception

See ForbiddenException

resource_limit_exceeded_exception

See ResourceLimitExceededException

service_failure_exception

See ServiceFailureException

service_unavailable_exception

See ServiceUnavailableException

throttled_client_exception

See ThrottledClientException

unauthorized_client_exception

See UnauthorizedClientException

Structs

BadRequestException

The input parameters don't match the service's restrictions.

ConflictException

The request could not be processed because of conflict in the current state of the

CreateAppInstanceAdminError

CreateAppInstanceError

CreateAppInstanceUserError

DeleteAppInstanceAdminError

DeleteAppInstanceError

DeleteAppInstanceUserError

DescribeAppInstanceAdminError

DescribeAppInstanceError

DescribeAppInstanceUserError

ForbiddenException

The client is permanently forbidden from making the request.

GetAppInstanceRetentionSettingsError

ListAppInstanceAdminsError

ListAppInstanceUsersError

ListAppInstancesError

PutAppInstanceRetentionSettingsError

ResourceLimitExceededException

The request exceeds the resource limit.

ServiceFailureException

The service encountered an unexpected error.

ServiceUnavailableException

The service is currently unavailable.

ThrottledClientException

The client exceeded its request rate limit.

UnauthorizedClientException

The client is not currently authorized to make the request.

UpdateAppInstanceError

UpdateAppInstanceUserError

Enums

CreateAppInstanceAdminErrorKind

CreateAppInstanceErrorKind

CreateAppInstanceUserErrorKind

DeleteAppInstanceAdminErrorKind

DeleteAppInstanceErrorKind

DeleteAppInstanceUserErrorKind

DescribeAppInstanceAdminErrorKind

DescribeAppInstanceErrorKind

DescribeAppInstanceUserErrorKind

GetAppInstanceRetentionSettingsErrorKind

ListAppInstanceAdminsErrorKind

ListAppInstanceUsersErrorKind

ListAppInstancesErrorKind

PutAppInstanceRetentionSettingsErrorKind

UpdateAppInstanceErrorKind

UpdateAppInstanceUserErrorKind