Module aws_sdk_chimesdkidentity::operation

Structs

CreateAppInstance

Creates an Amazon Chime SDK messaging AppInstance under an AWS account.

CreateAppInstanceAdmin

Promotes an AppInstanceUser to an AppInstanceAdmin. The

CreateAppInstanceUser

Creates a user under an Amazon Chime AppInstance. The request consists of a

DeleteAppInstance

Deletes an AppInstance and all associated data asynchronously.

DeleteAppInstanceAdmin

Demotes an AppInstanceAdmin to an AppInstanceUser. This action

DeleteAppInstanceUser

Deletes an AppInstanceUser.

DescribeAppInstance

Returns the full details of an AppInstance.

DescribeAppInstanceAdmin

Returns the full details of an AppInstanceAdmin.

DescribeAppInstanceUser

Returns the full details of an AppInstanceUser.

GetAppInstanceRetentionSettings

Gets the retention settings for an AppInstance.

ListAppInstanceAdmins

Returns a list of the administrators in the AppInstance.

ListAppInstanceUsers

List all AppInstanceUsers created under a single

ListAppInstances

Lists all Amazon Chime AppInstances created under a single AWS

PutAppInstanceRetentionSettings

Sets the amount of time in days that a given AppInstance retains

UpdateAppInstance

Updates AppInstance metadata.

UpdateAppInstanceUser

Updates the details of an AppInstanceUser. You can update names and